refactor: Move chart-loading-wrapper in charts directory
