+++ app/charts/bar/chart-bar.tsx
@@ -25,7 +25,7 @@ import {
 } from "@/graphql/query-hooks";
 import { useLocale } from "@/locales/use-locale";
 
-import { ChartLoadingWrapper } from "../scatterplot/chart-loading-wrapper";
+import { ChartLoadingWrapper } from "../chart-loading-wrapper";
 
 export const ChartBarsVisualization = ({
   dataSetIri,
