+++ app/charts/column/chart-column.tsx
@@ -37,7 +37,7 @@ import {
 } from "@/graphql/query-hooks";
 import { useLocale } from "@/locales/use-locale";
 
-import { ChartLoadingWrapper } from "../scatterplot/chart-loading-wrapper";
+import { ChartLoadingWrapper } from "../chart-loading-wrapper";
 
 export const ChartColumnsVisualization = ({
   dataSetIri,
