+++ app/charts/map/chart-map.tsx
@@ -32,7 +32,7 @@ import {
 } from "@/graphql/query-hooks";
 import { useLocale } from "@/locales/use-locale";
 
-import { ChartLoadingWrapper } from "../scatterplot/chart-loading-wrapper";
+import { ChartLoadingWrapper } from "../chart-loading-wrapper";
 
 export const ChartMapVisualization = ({
   dataSetIri,
