+++ app/charts/scatterplot/chart-scatterplot.tsx
@@ -32,7 +32,7 @@ import {
 } from "@/graphql/query-hooks";
 import { useLocale } from "@/locales/use-locale";
 
-import { ChartLoadingWrapper } from "./chart-loading-wrapper";
+import { ChartLoadingWrapper } from "../chart-loading-wrapper";
 
 export const ChartScatterplotVisualization = ({
   dataSetIri,
